modify : 숫자 계산 타입 long 으로 변환

+++ src/main/java/calculator/model/calculator/Calculator.java
@@ -1,5 +1,5 @@
 package calculator.model.calculator;
 
 public interface Calculator {
-    int calculate(int [] numbers);
+    long calculate(long [] numbers);
 }
